+++ e2e-tests/tests/angular-17.test.ts
@@ -195,9 +195,6 @@ function checkAngularProject(
       'replaysSessionSampleRate: 0.1',
       'replaysOnErrorSampleRate: 1',
       'enableLogs: true',
-      'dataCollection: {',
-      'userInfo: true,',
-      '}',
     ]);
   });