+++ e2e-tests/tests/angular-19.test.ts
@@ -194,9 +194,6 @@ function checkAngularProject(
       'replaysSessionSampleRate: 0.1',
       'replaysOnErrorSampleRate: 1',
       'enableLogs: true',
-      'dataCollection: {',
-      'userInfo: true,',
-      '}',
     ]);
   });