+++ e2e-tests/tests/react-native.test.ts
@@ -100,9 +100,9 @@ describe('ReactNative', () => {
 Sentry.init({
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
 
-  // Sends user identity info to Sentry. For more info visit:
+  // To disable sending user data, uncomment the line below. For more info visit:
   // https://docs.sentry.io/platforms/react-native/data-management/data-collected/
-  dataCollection: { userInfo: true },
+  // dataCollection: { userInfo: false },
 
   // Enable Logs
   enableLogs: true,