+++ e2e-tests/tests/sveltekit-hooks.test.ts
@@ -111,9 +111,9 @@ describe.sequential('Sveltekit', () => {
   // If you don't want to use Session Replay, just remove the line below:
   integrations: [replayIntegration()],
 
-  // Sends user identity info to Sentry. For more info visit:
+  // To disable sending user data, uncomment the line below. For more info visit:
   // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#dataCollection
-  dataCollection: { userInfo: true },
+  // dataCollection: { userInfo: false },
 });`,
         'export const handleError = handleErrorWithSentry(',
       ]);
@@ -131,9 +131,9 @@ describe.sequential('Sveltekit', () => {
   enableLogs: true,
 
 
-  // Sends user identity info to Sentry. For more info visit:
+  // To disable sending user data, uncomment the line below. For more info visit:
   // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#dataCollection
-  dataCollection: { userInfo: true },
+  // dataCollection: { userInfo: false },
 
   // uncomment the line below to enable Spotlight (https://spotlightjs.com)
   // spotlight: import.meta.env.DEV,
@@ -229,11 +229,7 @@ describe.sequential('Sveltekit', () => {
     replaysSessionSampleRate: 0.1,
     replaysOnErrorSampleRate: 1,
     integrations: [Sentry.replayIntegration()],
-    enableLogs: true,
-
-    dataCollection: {
-        userInfo: true
-    }
+    enableLogs: true
 })`,
         'export const handleError = Sentry.handleErrorWithSentry(',
       ]);
@@ -245,11 +241,7 @@ describe.sequential('Sveltekit', () => {
         `Sentry.init({
     dsn: "${TEST_ARGS.PROJECT_DSN}",
     tracesSampleRate: 1,
-    enableLogs: true,
-
-    dataCollection: {
-        userInfo: true
-    }
+    enableLogs: true
 })`,
         'export const handleError = Sentry.handleErrorWithSentry();',
       ]);