+++ e2e-tests/tests/sveltekit-tracing.test.ts
@@ -141,9 +141,9 @@ describe('Sveltekit with instrumentation and tracing', () => {
           // If you don't want to use Session Replay, just remove the line below:
           integrations: [replayIntegration()],
 
-          // Sends user identity info to Sentry. For more info visit:
+          // To disable sending user data, uncomment the line below. For more info visit:
           // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#dataCollection
-          dataCollection: { userInfo: true },
+          // dataCollection: { userInfo: false },
         });
 
         // If you have a custom error handler, pass it to \`handleErrorWithSentry\`