+++ src/react-router/codemods/client.entry.ts
@@ -47,7 +47,9 @@ const tracing = Sentry.reactRouterTracingIntegration({ useInstrumentationAPI: tr
 
 Sentry.init({
   dsn: "${dsn}",
-  dataCollection: { userInfo: true },
+  // To disable sending user data, uncomment the line below. For more info visit:
+  // https://docs.sentry.io/platforms/javascript/guides/react-router/configuration/options/#dataCollection
+  // dataCollection: { userInfo: false },
   integrations: [${integrations.join(', ')}],
   ${enableLogs ? 'enableLogs: true,' : ''}
   tracesSampleRate: 1.0,
@@ -69,7 +71,9 @@ Sentry.init({
       initContent = `
 Sentry.init({
   dsn: "${dsn}",
-  dataCollection: { userInfo: true },
+  // To disable sending user data, uncomment the line below. For more info visit:
+  // https://docs.sentry.io/platforms/javascript/guides/react-router/configuration/options/#dataCollection
+  // dataCollection: { userInfo: false },
   integrations: [${integrations.join(', ')}],
   ${enableLogs ? 'enableLogs: true,' : ''}
   tracesSampleRate: ${enableTracing ? '1.0' : '0'},${